Add a basic setup capable of running a couple instructions.
This includes a `Memory`, a `Cpu`, ability to load the bios from a file
and ability to execute a couple first instructions (LUI, ORI and SW).
Unfortunately, we hit a wall at SW, since writing to memory control
ports is not implemented.

This looks like a good start to me!